+++ lambdarpc-core/src/main/proto/Entity.proto
@@ -7,6 +7,8 @@ package io.lambdarpc.transport.grpc;
 
 import "FunctionPrototype.proto";
 
+// "Any" abject.
+// Function prototype exists to reduce serializations
 message Entity {
   oneof content {
     bytes data = 1;
