--- lambdarpc-core/src/main/proto/ExecuteError.proto
@@ -1,19 +1,0 @@
-syntax = "proto3";
-
-option java_multiple_files = true;
-option java_generic_services = true;
-option java_package = "io.lambdarpc.transport.grpc";
-package io.lambdarpc.transport.grpc;
-
-enum ErrorType {
-  OTHER = 0;
-  TIMEOUT_ERROR = 1;
-  SERIALIZATION_ERROR = 2;
-  WRONG_SERVICE_ERROR = 3;
-  FUNCTION_NOT_FOUND_ERROR = 4;
-}
-
-message ExecuteError {
-  ErrorType type = 1;
-  string message = 2;
-}
